requirements.txt
flask==3.1.2
pyyaml==6.0.2
flask-swagger-ui==5.21.0
openapi-core==0.19.5